+++ CHANGES.md
@@ -2,6 +2,9 @@
 
  * to be released
 
+ * A --reexport flag, which can be used to add extra modules to the
+   top-level module tree
+
  * Haddock no longer reports coverage statistics for hidden modules. By default
    cabal-install marks all package internal modules as hidden.
 
@@ -44,7 +47,7 @@
 
  * Remove framed view of the HTML documentation
 
-## Changes in version 2.17.2
+Changes in version 2.17.2
 
  * Fix portability of documentation building within GHC